Fold source set integration tester into test
